Tile pyramiding
Varies dramatically from the earlier version in chunk.py, as that
implementation resulting in large amounts of intermediate data sticking
around and needing to be shuffled between nodes.

This implementation ensures that merging occurs on a single node but
does not shuffle resulting data between nodes, preferring to reload it
from S3.